feat: add react-query for data fetching
