+++ package.json
@@ -11,6 +11,7 @@
   },
   "dependencies": {
     "lucide-react": "^0.511.0",
+    "@tanstack/react-query": "^5.0.0",
     "next": "15.3.2",
     "next-auth": "^4.24.10",
     "react": "^19.0.0",
