+++ src/app/api/accounts/route.ts
@@ -1,0 +1,94 @@
+import { NextResponse } from 'next/server';
+
+const accountsData = [
+  {
+    id: 1,
+    name: 'Primary Investment Portfolio',
+    type: 'Investment',
+    accountNumber: '****4521',
+    balance: 1850000,
+    change: 45000,
+    changePercent: 2.49,
+    assetTypes: ['Stocks', 'Bonds', 'ETFs'],
+    lastUpdated: '2025-05-29',
+    status: 'Active',
+    icon: 'TrendingUp',
+    color: 'text-blue-600',
+    bgColor: 'bg-blue-100',
+    riskLevel: 'Moderate',
+    ytdReturn: 12.4,
+  },
+  {
+    id: 2,
+    name: 'Retirement Savings Account',
+    type: 'Retirement',
+    accountNumber: '****8932',
+    balance: 650000,
+    change: -8500,
+    changePercent: -1.29,
+    assetTypes: ['Target Date Funds', 'Bonds'],
+    lastUpdated: '2025-05-29',
+    status: 'Active',
+    icon: 'Briefcase',
+    color: 'text-green-600',
+    bgColor: 'bg-green-100',
+    riskLevel: 'Conservative',
+    ytdReturn: 8.7,
+  },
+  {
+    id: 3,
+    name: 'Real Estate Investment Trust',
+    type: 'Real Estate',
+    accountNumber: '****2847',
+    balance: 250000,
+    change: 12000,
+    changePercent: 5.04,
+    assetTypes: ['REITs', 'Real Estate Funds'],
+    lastUpdated: '2025-05-28',
+    status: 'Active',
+    icon: 'Home',
+    color: 'text-orange-600',
+    bgColor: 'bg-orange-100',
+    riskLevel: 'Moderate-High',
+    ytdReturn: 15.2,
+  },
+  {
+    id: 4,
+    name: 'High-Yield Savings',
+    type: 'Savings',
+    accountNumber: '****5634',
+    balance: 75000,
+    change: 250,
+    changePercent: 0.33,
+    assetTypes: ['Cash', 'Money Market'],
+    lastUpdated: '2025-05-29',
+    status: 'Active',
+    icon: 'Banknote',
+    color: 'text-emerald-600',
+    bgColor: 'bg-emerald-100',
+    riskLevel: 'Low',
+    ytdReturn: 4.2,
+  },
+  {
+    id: 5,
+    name: 'Corporate Bond Portfolio',
+    type: 'Fixed Income',
+    accountNumber: '****9173',
+    balance: 180000,
+    change: 3200,
+    changePercent: 1.81,
+    assetTypes: ['Corporate Bonds', 'Treasury Bills'],
+    lastUpdated: '2025-05-29',
+    status: 'Active',
+    icon: 'Building2',
+    color: 'text-purple-600',
+    bgColor: 'bg-purple-100',
+    riskLevel: 'Low-Moderate',
+    ytdReturn: 6.8,
+  },
+];
+
+export async function GET() {
+  return NextResponse.json(accountsData);
+}
+
