+++ src/lib/queryClient.ts
@@ -1,0 +1,4 @@
+import { QueryClient } from '@tanstack/react-query';
+
+export const queryClient = new QueryClient();
+
